add clean flag and delete nodes when role controller works

+++ inheritance/naive_inheritance.sh
@@ -6,8 +6,9 @@
 #   https://github.com/openstack-k8s-operators/dataplane-operator/
 #   blob/main/controllers/suite_test.go
 
-FAKE_IT=1
+FAKE_IT=0
 VERBOSE=1
+CLEAN=1
 
 pushd /home/fultonj/zed/inheritance
 
@@ -21,9 +22,9 @@ function show_nodes() {
         echo "Zero nodes were created (but two should have been)"
     fi
     if [[ $VERBOSE -eq 1 ]]; then
-        # 2nd+3rd item of zero-indexed list (items[1]) should be sample-inheritance nodes
+        # 0th+1st item of zero-indexed list (items[1]) should be sample-inheritance nodes
+        oc get openstackdataplanenodes.dataplane.openstack.org -o json | jq .items[0]
         oc get openstackdataplanenodes.dataplane.openstack.org -o json | jq .items[1]
-        oc get openstackdataplanenodes.dataplane.openstack.org -o json | jq .items[2]
     fi
 }
 
@@ -51,9 +52,26 @@ if [[ $FAKE_IT -eq 1 ]]; then
     oc create -f node1.yaml
     oc create -f node2.yaml
     show_nodes
-    oc delete -f node1.yaml
-    oc delete -f node2.yaml
+    if [[ $CLEAN -eq 1 ]]; then
+        echo "Deleting after fake it"
+        oc delete -f node1.yaml
+        oc delete -f node2.yaml
+    fi
+else
+    if [[ $CLEAN -eq 1 ]]; then
+        NODES=$(oc get openstackdataplanenodes.dataplane.openstack.org \
+                    | grep inheritance | wc -l)
+        if [[ $NODES -gt 0 ]]; then
+            echo "Deleting nodes created by role"
+            # If we're not faking it and controller works, manually clean up
+            oc delete OpenStackDataPlaneNode dataplanenode-inheritance-0
+            oc delete OpenStackDataPlaneNode dataplanenode-inheritance-1
+        fi
+    fi
 fi
-
-oc delete -f role.yaml
+if [[ $CLEAN -eq 1 ]]; then
+    # Should this role deletion should also delete the nodes?
+    echo "Deleting role"
+    oc delete -f role.yaml
+fi
 popd
